Python/Containers/MongoDB-REST/requirements.txt
flask
prometheus_client
pymongo
python-dotenv
